separate module for value_generators
