+++ src/CMakeLists.txt
@@ -15,12 +15,13 @@
 add_subdirectory(third_party)
 
 add_subdirectory(metrics)
+add_subdirectory(value_generators)
 add_subdirectory(gennylib)
 add_subdirectory(cast_core)
 add_subdirectory(driver)
 
 # This makes the project importable from the build directory
-export(TARGETS metrics gennylib gennylib_cast gennylib_generators gennylib_core loki cast_core
+export(TARGETS value_generators metrics gennylib gennylib_cast gennylib_core loki cast_core
        FILE GennyLibraryConfig.cmake)
 
 # Below generates a cmake module so if you make install from this
