+++ src/cast_core/include/cast_core/actors/Loader.hpp
@@ -17,7 +17,7 @@
 
 
 #include <gennylib/Actor.hpp>
-#include <gennylib/DefaultRandom.hpp>
+#include <value_generators/value_generators.hpp>
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
 #include <metrics/metrics.hpp>
