+++ src/cast_core/include/cast_core/actors/MultiCollectionUpdate.hpp
@@ -16,7 +16,7 @@
 #define HEADER_D112CCC3_DF60_434E_A038_5A7AADED0E46
 
 #include <gennylib/Actor.hpp>
-#include <gennylib/DefaultRandom.hpp>
+#include <value_generators/value_generators.hpp>
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
 #include <metrics/metrics.hpp>
