+++ src/cast_core/src/actors/Insert.cpp
@@ -27,7 +27,7 @@
 #include <bsoncxx/json.hpp>
 #include <gennylib/Cast.hpp>
 #include <gennylib/context.hpp>
-#include <gennylib/value_generators.hpp>
+#include <value_generators/value_generators.hpp>
 
 namespace genny::actor {
