+++ src/cast_core/src/actors/InsertRemove.cpp
@@ -24,7 +24,7 @@
 
 #include <gennylib/Cast.hpp>
 #include <gennylib/context.hpp>
-#include <gennylib/value_generators.hpp>
+#include <value_generators/value_generators.hpp>
 
 namespace genny::actor {
