+++ src/cast_core/src/actors/MultiCollectionUpdate.cpp
@@ -28,7 +28,7 @@
 
 #include <gennylib/Cast.hpp>
 #include <gennylib/context.hpp>
-#include <gennylib/value_generators.hpp>
+#include <value_generators/value_generators.hpp>
 
 namespace genny::actor {
