+++ src/gennylib/CMakeLists.txt
@@ -39,6 +39,7 @@ target_link_libraries(gennylib_core
     yaml-cpp
     loki
     metrics
+    value_generators
     Boost::boost Boost::log
     MongoCxx::mongocxx
 )
@@ -53,19 +54,6 @@ add_library(gennylib_cast STATIC
 )
 target_link_libraries(gennylib_cast gennylib_core)
 
-# Concepts to do with generators
-add_library(gennylib_generators STATIC
-    src/value_generators/parser.cpp
-    src/value_generators/value_generators.cpp
-    src/value_generators/generators-private.cpp
-)
-
-target_include_directories(gennylib_generators
-    PRIVATE
-        src/value_generators
-)
-target_link_libraries(gennylib_generators gennylib_core)
-
 # Additional utilities and core concepts
 # Note that this contains both what you need to compile Actors and what you need to compile a Driver
 add_library(gennylib SHARED
@@ -73,7 +61,6 @@ add_library(gennylib SHARED
     src/context.cpp
     src/Orchestrator.cpp
     src/PoolFactory.cpp
-    src/DefaultRandom.cpp
     )
 
 # Define headers for this library. PUBLIC headers are used for
@@ -88,7 +75,6 @@ target_include_directories(gennylib
 target_link_libraries(gennylib
     PUBLIC
         gennylib_cast
-        gennylib_generators
         gennylib_core
 )
 if(ENABLE_JASPER)
@@ -107,7 +93,7 @@ target_link_libraries(gennylib
 )
 
 # 'make install' to the correct locations (provided by GNUInstallDirs).
-install(TARGETS  gennylib gennylib_cast gennylib_generators gennylib_core
+install(TARGETS  gennylib gennylib_cast gennylib_core
         EXPORT   GennyLibraryConfig
         ARCHIVE  DESTINATION ${CMAKE_INSTALL_LIBDIR}
         LIBRARY  DESTINATION ${CMAKE_INSTALL_LIBDIR}
@@ -141,8 +127,6 @@ add_executable(test_gennylib
     test/PhaseLoop_test.cpp
     test/version_test.cpp
     test/orchestrator_test.cpp
-    test/document_test.cpp
-    test/DefaultRandom_test.cpp
 
     # Test the helper
     test/ActorHelper_test.cpp
