+++ src/value_generators/test/document_test.cpp
@@ -23,10 +23,11 @@
 #include <bsoncxx/types.hpp>
 #include <bsoncxx/types/value.hpp>
 
-#include "../src/value_generators/generators-private.hh"
-#include "../src/value_generators/parser.hh"
-#include <gennylib/DefaultRandom.hpp>
-#include <gennylib/value_generators.hpp>
+#include "../src/generators-private.hh"
+#include "../src/parser.hh"
+
+#include <value_generators/value_generators.hpp>
+#include <value_generators/DefaultRandom.hpp>
 
 using namespace genny::value_generators;
 using bsoncxx::builder::stream::close_document;
